Don't set any specific minFrameRate for codecs
If no frame rate is specified for recording libmediaplayer
will set the frame rate to minFrameRate if that value is not -1, otherwise
it will use the Camera's default supported frame rate.

This can fail in CameraSource if that specific (min) frame rate is
not supported, and even if it is, it is preferable to set the default
frame rate to the Camera's default frame rate which can be more
than the codec's minFrameRate.

Issue: APDEV-949
Category: Aosp-Improvement
Domain: Aosp-media
Origin: internal
Upstream-Candidate: No
Change-Id: I3d5d163e45ea36b47128f617df1ed7bf099fd933
Signed-off-by: Adriana Reus <adriana.reus@intel.com>
Reviewed-on: https://android.intel.com/161742
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Chiras, Robert <robert.chiras@intel.com>
Reviewed-by: Breana, Tiberiu A <tiberiu.a.breana@intel.com>
Reviewed-by: Petrescu, Rares <rares.petrescu@intel.com>
Reviewed-by: Suman, Viorel <viorel.suman@intel.com>
Reviewed-by: Timohi, Alexandru <alexandru.timohi@intel.com>
Reviewed-by: Karlsson, Andreas <andreas.karlsson@intel.com>
Reviewed-by: Leung, Daniel <daniel.leung@intel.com>